requirements.txt
aiogram>=3.24.0
anthropic>=0.76.0
python-docx>=1.2.0
python-dotenv>=1.2.1
loguru>=0.7.3
pytest>=9.0.2
